+++ lib/que/scheduler.rb
@@ -1,2 +1,3 @@
 require 'que/scheduler/version'
 require 'que/scheduler/scheduler_job'
+require 'que/scheduler/engine' if Gem::Specification.find_by_name('railties')