+++ spec/spec_helper.rb
@@ -11,8 +11,7 @@ ENV['QUE_SCHEDULER_CONFIG_LOCATION'] = "#{__dir__}/config/que_schedule.yml"
 # Require zonebie before any other gem to ensure it sets the correct test timezone.
 require 'zonebie/rspec'
 
-require 'pry-byebug'
-require 'que/scheduler'
+Bundler.require :default, :test
 
 Dir["#{__dir__}/../spec/support/**/*.rb"].each { |f| require f }